GEM-18842: Update versions
- ben-manes-versions: 0.50.0 -> 0.54.0
- gradle-wrapper: 9.4.1 -> 9.5.1
- littlerobots-version-catalog-update: 0.8.5 -> 1.1.0
- logback-classic: 1.5.12 -> 1.5.33
- slf4j: 2.0.16 -> 2.0.18
- spring-framework-bom: 6.2.12 -> 6.2.18
- spring-security-bom: 6.5.6 -> 6.5.10
- spring-session-core: 3.5.3 -> 3.5.6

ai-assisted=yes
