+++ build-tools/convention-plugins/build.gradle.kts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 Broadcom. All rights reserved.
+ * Copyright 2024-2026 Broadcom. All rights reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
 
@@ -9,8 +9,10 @@ plugins {
 }
 
 repositories {
-  mavenCentral()
-  gradlePluginPortal()
+  addGemFireRepositories(
+    providers,
+    addMavenCentral = providers.gradleProperty("useMavenCentral").getOrElse("false").toBoolean()
+  )
 }
 
 dependencies {
@@ -24,4 +26,30 @@ gradlePlugin {
     id = "gemfire-artifactory"
     implementationClass = "com.vmware.gemfire.gradle.ArtifactoryPlugin"
   }
+}
+
+fun RepositoryHandler.addGemFireRepositories(
+  providers: ProviderFactory,
+  addGradlePluginPortal: Boolean = false,
+  addMavenCentral: Boolean = false
+) {
+  val configFilePath = providers.gradleProperty("spring.gemfire.repositories").getOrElse(
+    providers.environmentVariable("HOME").get() + "/.gradle/gradleRepositories.json"
+  )
+  val jsonString = File(configFilePath).readText(Charsets.UTF_8)
+  val repos = groovy.json.JsonSlurper().parseText(jsonString) as Map<*, *>
+  (repos["repositories"] as List<*>).filterNotNull().map { it as Map<*, *> }
+    .forEach { entry ->
+      maven {
+        url = uri(entry["url"]!! as String)
+        if (!entry["username"]?.toString().isNullOrBlank()) {
+          credentials {
+            username = entry["username"] as String
+            password = entry["password"] as String
+          }
+        }
+      }
+    }
+  if (addGradlePluginPortal) gradlePluginPortal()
+  if (addMavenCentral) mavenCentral()
 }
