+++ build-tools/convention-plugins/src/main/kotlin/commercial-repositories.gradle.kts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 Broadcom. All rights reserved.
+ * Copyright 2024-2026 Broadcom. All rights reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
 
@@ -10,34 +10,5 @@ plugins {
 }
 
 repositories {
-  mavenCentral()
-  val repositoryConfigFilePath = providers.gradleProperty("spring.gemfire.repositories").getOrElse(
-    providers.environmentVariable("HOME").get() + "/.gradle/gradleRepositories.json"
-  )
-
-  val enablePrivateCommercialRepos =
-    providers.gradleProperty("spring.gemfire.enable.private.repositories")
-      .getOrElse("false").toBoolean()
-  val jsonString = File(repositoryConfigFilePath).readText(Charsets.UTF_8)
-  val repositories = groovy.json.JsonSlurper().parseText(jsonString) as Map<*, *>
-  (repositories["repositories"] as List<*>).filterNotNull().map { entry -> entry as Map<*, *> }
-    .filter { entry ->
-      return@filter if (entry["private"]!! as Boolean) {
-        enablePrivateCommercialRepos
-      } else {
-        true
-      }
-    }.forEach { entry ->
-      entry.apply {
-        maven {
-          url = uri(entry["url"]!! as String)
-          if (!entry["username"]?.toString().isNullOrBlank()) {
-            credentials {
-              username = entry["username"] as String
-              password = entry["password"] as String
-            }
-          }
-        }
-      }
-    }
+  addGemFireRepositories(providers)
 }
