+++ gradle/libs.versions.toml
@@ -1,23 +1,25 @@
 [versions]
 assertjVersion = "3.23.1"
 awaitilityVersion = "4.2.2"
-benManesVersions = "0.50.0"
+benManesVersions = "0.54.0"
 gemFireTestContainersVersion = "3.0.0"
+gemfireVersion = "10.1.7"
 jakartaAnnotationApiVersion = "2.1.1"
 jakartaServletApiVersion = "6.0.0"
 jsr305Version = "3.0.2"
 junitVersion = "4.13.2"
-littlerobotsVersionCatalogUpdate = "0.8.5"
-logbackVersion = "1.5.12"
+littlerobotsVersionCatalogUpdate = "1.1.0"
+logbackVersion = "1.5.33"
 lombokVersion = "8.6"
 mockitoVersion = "5.14.2"
 multithreadedtcVersion = "1.01"
 nebulaFacetVersion = "11.0.0"
-slf4jVersion = "2.0.16"
-springSecurityVersion = "6.5.6"
-springSessionVersion = "3.5.3"
+slf4jVersion = "2.0.18"
+springDataGemFireVersion = "2.0.1"
+springSecurityVersion = "6.5.10"
+springSessionVersion = "3.5.6"
 springShellVersion = "1.2.0.RELEASE"
-springVersion = "6.2.12"
+springVersion = "6.2.18"
 
 [libraries]
 assertj-core = { module = "org.assertj:assertj-core", version.ref = "assertjVersion" }
